+++ lib/qualtrics_api/panel.rb
@@ -1,19 +1,10 @@
 module QualtricsAPI
-  class Panel
-    include Virtus.value_object
-    include QualtricsAPI::Extensions::SerializableModel
+  class Panel < BaseModel
 
     attribute :id, String
     attribute :library_id, String
     attribute :name, String
     attribute :category, String
-
-    def initialize(options = {})
-      attributes_mappings.each do |key, qualtrics_key|
-        instance_variable_set "@#{key}", options[qualtrics_key]
-      end
-      super
-    end
 
     def members(options = {})
       @members ||= QualtricsAPI::PanelMemberCollection.new(options.merge(id: id))
